+++ inst/bs4Dash-0.2.0/update-tabs.js
@@ -35,6 +35,36 @@ $(function() {
     });
   });
   
+  // handles the insertTab function
+  tabIds.forEach(function(index) {
+    var id = "insert_" + index;
+    Shiny.addCustomMessageHandler(id, function(message) {
+      var tabId = message.ns + "-" + message.target;
+      if (message.position === "after") {
+        // insert after the targeted tag in the tab-panel div
+        $(message.value).insertAfter($("#" +tabId));
+        // we also need to insert an item in the navigation
+        $(message.link).insertAfter($('[href ="#' + tabId + '"]').parent());
+      } else if (message.position === "before") {
+        // insert before the targeted tag in the tab-panel div
+        $(message.value).insertBefore($("#" + tabId));
+        // we also need to insert an item in the navigation
+        $(message.link).insertBefore($('[href ="#' + tabId + '"]').parent());
+      }
+      
+      // if the newly inserted tab is active, disable other tabs
+      if (message.select === "true") {
+        // trigger a click on corresponding the new tab button. 
+        $('#' + index + ' a[href="#' +  message.id +'"]').trigger('click');
+        $('#' + message.id).addClass('active show');
+        
+        // remove active class from ul menu + the body content (tab content)
+        $('#' + index + ' a[href!="#' + message.id +'"]').removeClass('active');
+        $('div[id="' +  message.id + '"]').siblings().removeClass('active show');
+      }
+    });
+  });
+  
   
   // handle the right controlbar 
   var controlbarIds = [];